+++ templates/pages/indexLayout.html.twig
@@ -1,0 +1,7 @@
+{% extends 'base.html.twig' %}
+
+{% block hero %}
+{% endblock %}
+
+{% block content %}
+{% endblock %}
